separate internal-tests and unit-tests projects
