+++ unit-tests/unit-tests-expected.h
@@ -18,21 +18,6 @@
 #include "ds5/ds5-device.h"
 #include "frame-validator.h"
 
-typedef uint16_t PID_t;
-
-const std::string L500_TAG = "L500";
-const std::string L500_COLOR_SENSOR_NAME = "RGB Camera";
-const std::string L500_DEPTH_SENSOR_NAME = "L500 Depth Sensor";
-const std::string L500_MOTION_SENSOR_NAME = "Motion Module";
-
-inline PID_t string_to_hex(std::string pid)
-{
-    std::stringstream str(pid);
-    PID_t value = 0;
-    str >> std::hex >> value;
-    return value;
-}
-
 // Builds profiles by desired configurations
 struct expected_profiles
 {
@@ -494,174 +479,6 @@ inline std::unordered_map<std::string, std::vector<rs2_frame_metadata_value>> ge
     PID_t pid_num = string_to_hex(pid);
 
     return generate_device_metadata(pid_num);
-}
-
-inline std::unordered_map<std::string,
-    std::vector<std::pair<std::vector<profile>, std::vector<profile>>>> generate_sensor_resolver_profiles(std::string pid)
-{
-    // pair first = source profiles to resolve
-    // pair second = resolved target profiles
-
-    using namespace librealsense;
-    using namespace ds;
-
-    PID_t pid_num = string_to_hex(pid);
-
-    switch (pid_num)
-    {
-    case L500_PID:
-    case L515_PID:
-        return
-        {
-            {
-                // DEPTH SENSOR
-                L500_DEPTH_SENSOR_NAME,
-                { // vector of expected pairs
-                    { // pairs of source profiles to resolve and expected resolved profiles
-                        { // vector of source profiles to resolve
-                            { RS2_STREAM_DEPTH,      RS2_FORMAT_Z16,  0, 640, 480, 30 },
-                            { RS2_STREAM_INFRARED,   RS2_FORMAT_Y8,   0, 640, 480, 30 },
-                            { RS2_STREAM_CONFIDENCE, RS2_FORMAT_RAW8, 0, 640, 480, 30 },
-                        },
-                        { // vector of expected resolved target profiles
-                            { RS2_STREAM_DEPTH,      RS2_FORMAT_Z16,  0, 640, 480, 30 },
-                            { RS2_STREAM_INFRARED,   RS2_FORMAT_Y8,   0, 640, 480, 30 },
-                            { RS2_STREAM_CONFIDENCE, RS2_FORMAT_RAW8, 0, 640, 480, 30 },
-                        },
-                    },
-                },
-            },
-            {
-                // COLOR SENSOR
-                L500_COLOR_SENSOR_NAME,
-                { // vector of expected pairs
-                    { // pairs of source profiles to resolve and expected resolved profiles
-                        { // vector of source profiles to resolve
-                            { RS2_STREAM_COLOR,      RS2_FORMAT_RGB8,  0, 640, 480, 30 },
-                        },
-                        { // vector of expected resolved target profiles
-                            { RS2_STREAM_COLOR,      RS2_FORMAT_YUYV,  0, 640, 480, 30 },
-                        },
-                    },
-                    {
-                        {
-                            { RS2_STREAM_COLOR,      RS2_FORMAT_YUYV,  0, 640, 480, 30 },
-                        },
-                        {
-                            { RS2_STREAM_COLOR,      RS2_FORMAT_YUYV,  0, 640, 480, 30 },
-                        },
-                    },
-                },
-            },
-            {
-                // COLOR SENSOR
-                L500_MOTION_SENSOR_NAME,
-                { // vector of expected pairs
-                    { // pairs of source profiles to resolve and expected resolved profiles
-                        { // vector of source profiles to resolve
-                            { RS2_STREAM_ACCEL,      RS2_FORMAT_MOTION_XYZ32F,  0, 0, 0, 100 },
-                        },
-                        { // vector of expected resolved target profiles
-                            { RS2_STREAM_ACCEL,      RS2_FORMAT_MOTION_XYZ32F,  0, 0, 0, 100 },
-                        },
-                    },
-                    {
-                        {
-                            { RS2_STREAM_GYRO,      RS2_FORMAT_MOTION_XYZ32F,  0, 0, 0, 100 },
-                        },
-                        {
-                            { RS2_STREAM_GYRO,      RS2_FORMAT_MOTION_XYZ32F,  0, 0, 0, 100 },
-                        },
-                    },
-                },
-            }
-        };
-    default:
-        WARN("Device does not have expected data.");
-        return {};
-    }
-}
-
-inline std::map<std::string, std::map<uint32_t, rs2_format>> generate_fourcc_to_rs2_format_map(std::string pid)
-{
-    using namespace librealsense;
-    using namespace ds;
-
-    PID_t pid_num = string_to_hex(pid);
-
-    switch (pid_num)
-    {
-    case L500_PID:
-    case L515_PID:
-        return
-        {
-            {
-                L500_DEPTH_SENSOR_NAME,
-                {
-                    { rs_fourcc('G','R','E','Y'), RS2_FORMAT_Y8 },
-                    { rs_fourcc('Z','1','6',' '), RS2_FORMAT_Z16 },
-                    { rs_fourcc('C',' ',' ',' '), RS2_FORMAT_RAW8 },
-                }
-            },
-            {
-                L500_COLOR_SENSOR_NAME,
-                {
-                    {rs_fourcc('Y','U','Y','2'), RS2_FORMAT_YUYV},
-                    {rs_fourcc('Y','U','Y','V'), RS2_FORMAT_YUYV},
-                    {rs_fourcc('U','Y','V','Y'), RS2_FORMAT_UYVY}
-                }
-            },
-            {
-                L500_MOTION_SENSOR_NAME,
-                {
-                }
-            },
-        };
-    default:
-        WARN("Device does not have expected data.");
-        return {};
-    }
-}
-
-inline std::map<std::string, std::map<uint32_t, rs2_stream>> generate_fourcc_to_rs2_stream_map(std::string pid)
-{
-    using namespace librealsense;
-    using namespace ds;
-
-    PID_t pid_num = string_to_hex(pid);
-
-    switch (pid_num)
-    {
-    case L500_PID:
-    case L515_PID:
-        return
-        {
-            {
-                L500_DEPTH_SENSOR_NAME,
-                {
-                    { rs_fourcc('G','R','E','Y'), RS2_STREAM_INFRARED },
-                    { rs_fourcc('Z','1','6',' '), RS2_STREAM_DEPTH },
-                    { rs_fourcc('C',' ',' ',' '), RS2_STREAM_CONFIDENCE }
-                }
-            },
-            {
-                L500_COLOR_SENSOR_NAME,
-                {
-                    {rs_fourcc('Y','U','Y','2'), RS2_STREAM_COLOR},
-                    {rs_fourcc('Y','U','Y','V'), RS2_STREAM_COLOR},
-                    {rs_fourcc('U','Y','V','Y'), RS2_STREAM_COLOR}
-                }
-            },
-            {
-                L500_MOTION_SENSOR_NAME,
-                {
-                }
-            },
-        };
-    default:
-        WARN("Device does not have expected data.");
-        return {};
-    }
 }
 
 #endif //LIBREALSENSE_UNITTESTS_GROUNDTRUTH_H
